Added UNDERTALE and Pillars of Eternity (native linux) to database.json (#48)
* Added UNDERTALE and Pillars of Eternity (native linux) to database.json
Note that Pillars behaves oddly on launch; it opens and closes a few
times due to how it handles the intro video. This is fixable by delaying
the usage of fix-wm-class.sh. I achieved this by changing the launch
options to `%command% & sleep 5 && fix-wm-class.sh "Pillars of Eternity";`
If this is not acceptable, I can perhaps find a way to work that into the
actual sif.sh script. Just let me know.

* Check if Pillars of Eternity
 If so, add & `sleep 5` && to Launch Options, after existing commands
 but before fix-wm-class.sh

* Sorted database.json